Fix float64 leaks when running with precision=1
Several arrays, constants and numpy float64 scalars were created without
the object's dtype, so float32 simulations silently ran parts of the
pipeline (and some output data objects) in float64/complex128.

- processing_objects: cast setup-time arrays (SH slopec weights, DM m2c
  and stroke, atmo propagators, modulation/LIFT tilts, filter states,
  gain vectors, operation constants, ...) to self.dtype/complex_dtype;
  convert float64 scalars used per step to python floats; pass precision
  to internally created data objects.
- data_objects: infinite phase screen matrices and screens, Recmat/Intmat
  restored from FITS, masks, and helper methods now honour self.dtype.
- lib: masks returned by ifunc generators use the requested dtype;
  demodulate_signal honours dtype; EF interpolator avoids float64 mask.
- BaseValue.restore casts float arrays to the object dtype.
- cur_wfs_slopec no longer forces float32 when precision=0.

Random draws are kept in float64 and cast, so seeded sequences are
unchanged.

+++ specula/data_objects/ifunc.py
@@ -85,7 +85,7 @@ class IFunc(BaseDataObj):
                 raise ValueError(f'Invalid ifunc type {type_str}')
 
         ifunc = self.to_xp(ifunc)
-        mask = self.to_xp(mask)
+        mask = self.to_xp(mask, dtype=self.dtype)
 
         self._influence_function = ifunc
         self._mask_inf_func = mask
