+++ specula/processing_objects/poly_chrom_wfs.py
@@ -32,7 +32,7 @@ class PolyChromWFS(BaseProcessingObj):
             raise ValueError("xy_tilts_in_arcsec must have the same length as wavelengthInNm")
 
         self.wavelengths_in_nm = wavelengthInNm
-        self.flux_factor = self.to_xp(flux_factor)
+        self.flux_factor = self.to_xp(flux_factor, dtype=self.dtype)
         self.xy_tilts_in_arcsec = xy_tilts_in_arcsec
         self.n_wavelengths = n_wavelengths
 
@@ -60,7 +60,7 @@ class PolyChromWFS(BaseProcessingObj):
     def _create_unit_tilts(self, in_ef_size, in_ef_pixel_pitch):
         """Create unit tilt phase arrays (1 pixel tilt) in nm."""
         # Create coordinate arrays in meters
-        xx, yy = make_xy(in_ef_size, in_ef_pixel_pitch, xp=self.xp)
+        xx, yy = make_xy(in_ef_size, in_ef_pixel_pitch, xp=self.xp, dtype=self.dtype)
 
         # Calculate pupil diameter in meters
         pupil_diameter = in_ef_size * in_ef_pixel_pitch
